added test to kill pod
