+++ agent/Dockerfile
@@ -26,7 +26,7 @@ RUN curl -LO http://downloads.mongodb.org/linux/mongodb-linux-x86_64-ubuntu1604-
     mv mongodb-linux-x86_64-ubuntu1604-4.0.6/bin/mongo /usr/bin && \
     rm -rf mongodb-linux-x86_64-ubuntu1604-4.0.6.tgz mongodb-linux-x86_64-ubuntu1604-4.0.6
 
-RUN curl https://readinessprobe.s3-us-west-1.amazonaws.com/readinessprobe -o /agent/readinessprobe \
+RUN curl --retry 3 https://readinessprobe.s3-us-west-1.amazonaws.com/readinessprobe -o /agent/readinessprobe \
     && mkdir -p /var/log/mongodb-mms-automation/ \
     && chmod +x /agent/readinessprobe
