--- test/e2e/main_test.go
@@ -1,11 +1,0 @@
-package e2e
-
-import (
-	"testing"
-
-	f "github.com/operator-framework/operator-sdk/pkg/test"
-)
-
-func TestMain(m *testing.M) {
-	f.MainEntry(m)
-}
